- added moving object class

+++ MyGame/src/CollisionHandler/CollisionHandler.cpp
@@ -15,11 +15,11 @@ CollisionHandler* CollisionHandler::GetInstance() {
 }
 //horizontal collision handler
 bool CollisionHandler::CheckMapCollision(ObjProperty objProperty) {
-	SDL_Rect objectRect = { objProperty.xPosition ,objProperty.yPosition , objProperty.width, objProperty.height };
+	SDL_Rect objectRect = { (int)objProperty.xPosition ,(int)objProperty.yPosition , objProperty.width, objProperty.height };
 	tileMapList = MapParser::GetInstance()->GetMapLayers();
 	for (unsigned int t = 0;t < tileMapList.size();t++) {
-		for (unsigned int row = 0;row < tileMapList[t].rowCount;row++) {
-			for (unsigned int col = 0;col < tileMapList[t].colCount;col++) {
+		for (int row = 0;row < tileMapList[t].rowCount;row++) {
+			for (int col = 0;col < tileMapList[t].colCount;col++) {
 				bool isIgnoreTile = std::find(std::begin(ignoreTile), std::end(ignoreTile), tileMapList[t].tileMap[row][col]) != std::end(ignoreTile);
 				if (!isIgnoreTile) {
 					SDL_Rect tileRect = { row*tileMapList[t].tileWidth, col* tileMapList[t].tileHeight, tileMapList[t].tileWidth-1, tileMapList[t].tileHeight-1 };
@@ -35,11 +35,11 @@ bool CollisionHandler::CheckMapCollision(ObjProperty objProperty) {
 }
 //vertical collision handler
 bool CollisionHandler::CheckMapCollision(ObjProperty objProperty,int y) {
-	SDL_Rect objectRect = { objProperty.xPosition ,objProperty.yPosition , objProperty.width, objProperty.height };
+	SDL_Rect objectRect = { (int)objProperty.xPosition ,(int)objProperty.yPosition , objProperty.width, objProperty.height };
 	tileMapList = MapParser::GetInstance()->GetMapLayers();
 	for (unsigned int t = 0;t < tileMapList.size();t++) {
-		for (unsigned int row = 0;row < tileMapList[t].rowCount;row++) {
-			for (unsigned int col = 0;col < tileMapList[t].colCount;col++) {
+		for (int row = 0;row < tileMapList[t].rowCount;row++) {
+			for (int col = 0;col < tileMapList[t].colCount;col++) {
 				if (tileMapList[t].tileMap[row][col] != 0) {
 					SDL_Rect tileRect = { row * tileMapList[t].tileWidth, col * tileMapList[t].tileHeight, tileMapList[t].tileWidth - 1, tileMapList[t].tileHeight - 1 };
 					bool isIgnoreTile = std::find(std::begin(ignoreTile), std::end(ignoreTile), tileMapList[t].tileMap[row][col]) != std::end(ignoreTile);
